Add processing logic of error handling which is sending the error logs to the LLM and retry

+++ src/schemas/virality_score.py
@@ -7,7 +7,7 @@ class ScoreBreakdown(BaseModel):
     comment_potential: int
     execution_barrier: int
     compliance_safety: int
-    memory_fit_score: int
+    memory_fit_score: float
     
 class ScoreResult(BaseModel):
     total_score: float
